added Mod window
