+++ mods/__init__.py
@@ -1,17 +1,6 @@
 import os
 import importlib
-
-
-class ModError(Exception):
-    """
-    An Error that is thrown, when an inproper module is loaded.
-    """
-
-    def __init__(self, name, err):
-        self.name = name
-        self.err = err
-        super().__init__(f"The mod '{name}' lacks attributes!\n{err}")
-
+from pokete_classes.mods import ModError
 
 mod_names = [i.strip(".py") for i in os.listdir(__file__.strip("__init__.py"))
              if i[0] != "_"]
